Rename EmitConstants to emit_constants
