+++ examples/07-EMIT/ComputeProtectionLevels.py.back
@@ -17,7 +17,7 @@ import sys
 import subprocess
 import pyaedt
 from pyaedt import Emit
-from pyaedt.emit_core.EmitConstants import TxRxMode, ResultType, InterfererType
+from pyaedt.emit_core.emit_constants import TxRxMode, ResultType, InterfererType
 
 # Check to see which Python libraries have been installed
 reqs = subprocess.check_output([sys.executable, '-m', 'pip', 'freeze'])
