+++ pyaedt/emit_core/emit_constants.py
@@ -2,7 +2,7 @@ from pyaedt import emit_core
 
 """
 Enums from EmitApiPython are defined as none until the EmitApiPython module initializes.
-This allows IDE auto-complete to find them and EmitConstants to import before the
+This allows IDE auto-complete to find them and emit_constants to import before the
 EmitApiPython module has loaded (typically when a pyaedt.Emit() object is created).
 """
