+++ pyaedt/modeler/circuits/PrimitivesEmit.py
@@ -1,6 +1,6 @@
 from collections import defaultdict
 
-from pyaedt.emit_core import EmitConstants as emit_consts
+from pyaedt.emit_core import emit_constants as emit_consts
 import pyaedt.generic.constants as consts
 
 # from pyaedt.generic.general_methods import property
